one more proof and some documentation
